"Modified config_alpha function in the build script: updated location of libblas.so to match its new location after upgrading to Ubuntu 18.04 aka Bionic Beaver.
Modified the petsc section of configure.in:
now follows the same pattern used for CGAL to specify the include and
library directories seperately. This makes it easier to link from a
package that was installed by the package manager in the current format,
but more importantly allows us to start making better use of the package
manager in the short term. My goal is to have all external packages used
by FronTier to interface through the package manager and implement a
modern build system with either GNU Autotools or CMake (potentially both).